Scope dependency cache to full packageSourceMapping rules; split dependency tests
Copilot re-review round 13 flagged that ConfigScopeKey recorded only whether
<packageSourceMapping> was enabled, not the actual mapping rules. Two nuget.config
files with identical sources and globalPackagesFolder but different package-to-source
mappings (e.g. *->x vs *->y) produced byte-identical scope keys and could share the
process-wide dependency cache, so a dependency resolved from one feed under config A
could be served from cache to config B (which maps it to a different feed).

- NugetSourceProvider.ConfigScopeKey now fingerprints the complete mapping entries
  (source key -> ordered patterns) via PackageSourceMappingProvider.GetPackageSourceMappingItems(),
  ordered stably, instead of PackageSourceMapping.IsEnabled.
- Add regression test GetPackageDependenciesAsync_SameSourcesDifferentMapping_DoNotShareCache.
- Extract the dependency-resolution + cache-scoping tests into a new
  NugetServiceDependencyTests.cs to keep NugetServiceFeedTests.cs under the file-size limit.
